% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bin.R
\name{build_bin}
\alias{build_bin}
\title{build GRAN binaries}
\usage{
build_bin(sync = FALSE, GRAN.dir = "./GRAN")
}
\arguments{
\item{sync}{use S3 sync with packages?}

\item{GRAN.dir}{local directory for GRAN built packages}
}
\description{
build binary packages for GRAN from source
}

