% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_GRAN_all.R
\name{build_GRAN_all}
\alias{build_GRAN_all}
\title{build all src and binaries for specified versions of R}
\usage{
build_GRAN_all(versions)
}
\arguments{
\item{versions}{character vector of R major.minor versions (e.g., '3.2').}
}
\description{
builds binaries using \code{\link{dl_build_src}} and \code{\link{dl_build_bin}} for
each of the versions specified with the \code{versions} parameter.
}
\examples{
\dontrun{
build_GRAN_all(c('3.2','3.1'))
}
}

