% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bin.R
\name{checkAvailablePackages}
\alias{checkAvailablePackages}
\title{compare available packages in two directories}
\usage{
checkAvailablePackages(src, bin)
}
\arguments{
\item{src}{character path to source directory}

\item{bin}{character path to binary directory}
}
\description{
compare available packages in two directories
}

