% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_src_list.R
\name{checkVs}
\alias{checkVs}
\title{checks if "v" is appended in a column of version numbers, and adds it if it is not}
\usage{
checkVs(input)
}
\arguments{
\item{input}{character}
}
\description{
checks if "v" is appended in a column of version numbers, and adds it if it is not
}

