% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dep_order.R
\name{collapseDepGraph}
\alias{collapseDepGraph}
\title{Collapse dependency graph into list ordered from outside inward}
\usage{
collapseDepGraph(graph)
}
\arguments{
\item{graph}{igraph object to collapse}
}
\value{
list representing order in which to install dependencies
}
\description{
Collapse dependency graph into list ordered from outside inward
}
