% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compile_gran.R
\name{dl_build_bin}
\alias{dl_build_bin}
\title{build GRAN binaries}
\usage{
dl_build_bin(sync = FALSE)
}
\arguments{
\item{sync}{use S3 sync with packages?}
}
\description{
build binary packages for GRAN from source
}

