% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_src_list.R
\name{findNotMatched}
\alias{findNotMatched}
\title{findNotMatched}
\usage{
findNotMatched(x.1, x.2)
}
\arguments{
\item{x.1}{something}

\item{x.2}{something else}
}
\description{
checks two data frames and returns rows in 1 that aren't matched in 2
modified from \url{https://rappster.wordpress.com/2011/10/12/identifying-records-in-data-frame-a-that-are-not-contained-in-fata-frame-b-a-comparison}
}

