% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_gran_html.R
\name{generate_gran_html}
\alias{generate_gran_html}
\title{Generate GRAN package index and html sites}
\usage{
generate_gran_html(GRAN.dir = "./GRAN")
}
\description{
There are many ways to do this:
1. \url{http://watson.nci.nih.gov/bioc_mirror/packages/release/bioc/vignettes/biocViews/inst/doc/createReposHtml.pdf}|
2. https://github.com/hadley/staticdocs
3. Rd2HTML (maybe?)
4. install.packages('biocViews', repos = c('http://bioconductor.org/packages/release/bioc', 'http://cran.rstudio.com'))
}
