% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dep_order.R
\name{makeDepGraphFromDescription}
\alias{makeDepGraphFromDescription}
\title{Borrows from makeDepGraph but works on package not yet built}
\usage{
makeDepGraphFromDescription(
  file,
  suggests = TRUE,
  enhances = FALSE,
  includeBasePkgs = FALSE
)
}
\arguments{
\item{file}{DESCRIPTION file path}

\item{suggests}{include suggests}

\item{enhances}{include enhances}

\item{includeBasePkgs}{include base packges}
}
\description{
Borrows from makeDepGraph but works on package not yet built
}
