% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dir_version.R
\name{set_version}
\alias{set_version}
\title{set directory version of R}
\usage{
set_version(version)
}
\arguments{
\item{version}{the major.minor version to use}
}
\description{
Set version of R as major.minor for current. As character.
}
\examples{
\dontrun{
dir_version()
set_version('3.3')
}
}

