% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_build_src_gran.R
\name{writeBuildList}
\alias{writeBuildList}
\title{generate a build list file in a directory}
\usage{
writeBuildList(path)
}
\arguments{
\item{path}{character path to the folder to write buildList in}
}
\description{
generate a build list file in a directory
}
