% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dl_build_bin.R
\name{dl_build_bin}
\alias{dl_build_bin}
\title{build GRAN binaries}
\usage{
dl_build_bin(sync = FALSE, GRAN.dir = "../GRAN")
}
\arguments{
\item{sync}{use S3 sync with packages?}

\item{GRAN.dir}{local directory for GRAN built packages}
}
\description{
build binary packages for GRAN from source
}

