% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_build_src_gran.R
\name{dl_build_src}
\alias{dl_build_src}
\title{Build packages w/ deps}
\usage{
dl_build_src(GRAN.dir = "GRAN", lib = .libPaths()[1])
}
\arguments{
\item{GRAN.dir}{local directory for GRAN built packages}

\item{lib}{local directory to install dependent packages}
}
\description{
Build GRAN packages w/ GRAN and CRAN deps
}
