% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dl_build_src_gran.R
\name{dl_build_src}
\alias{dl_build_src}
\title{download and build packages w/ deps}
\usage{
dl_build_src(GRAN.dir = "../GRAN")
}
\arguments{
\item{GRAN.dir}{local directory for GRAN built packages}
}
\description{
Download and build GRAN packages w/ GRAN and CRAN deps
}

